# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API que implementa a cifra de Vernam (criptografia e descriptografia de mensagens).",
        "title": "API Cifra de Vernam",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/cifrar": {
            "post": {
                "description": "Aplica a cifra de Vernam ao texto fornecido com a chave informada.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "criptografia"
                ],
                "summary": "Cifrar uma mensagem com uma chave",
                "parameters": [
                    {
                        "description": "Texto e chave",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RequestCifrarDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseCifrarDTO"
                        }
                    }
                }
            }
        },
        "/decifrar": {
            "post": {
                "description": "Aplica a cifra de Vernam ao texto fornecido com a chave informada.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "descriptografia"
                ],
                "summary": "Decifrar mensagem com cifra de vernan",
                "parameters": [
                    {
                        "description": "Texto e chave",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RequestDecifrarDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseDecifrarDTO"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.RequestCifrarDTO": {
            "type": "object",
            "required": [
                "chave",
                "textoClaro"
            ],
            "properties": {
                "chave": {
                    "type": "string"
                },
                "textoClaro": {
                    "type": "string"
                }
            }
        },
        "models.RequestDecifrarDTO": {
            "type": "object",
            "properties": {
                "chave": {
                    "type": "string"
                },
                "textoCifrado": {
                    "type": "string"
                }
            }
        },
        "models.ResponseCifrarDTO": {
            "type": "object",
            "properties": {
                "textoCifrado": {
                    "type": "string"
                }
            }
        },
        "models.ResponseDecifrarDTO": {
            "type": "object",
            "required": [
                "textoClaro"
            ],
            "properties": {
                "textoClaro": {
                    "type": "string"
                }
            }
        }
    }
}